cassandra-spark-project/Django-Backend/myproject/requirements.txt
apt-clone==0.2.1
apturl==0.5.2
argcomplete==1.8.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
backports.tarfile==1.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
Brlapi==0.8.3
Brotli==1.1.0
cassandra-driver==3.29.2
cassandra-pylib==0.0.0
catfish==4.16.4
ccsm==0.9.14.1
certifi==2024.7.4
cffi==1.16.0
channels==4.1.0
chardet==5.2.0
charset-normalizer==3.3.2
cinemagoer==2023.5.1
click==8.1.7
colorama==0.4.6
command-not-found==0.3
compizconfig-python==0.9.14.1
configobj==5.0.8
constantly==23.10.4
coverage==5.5
cryptography==43.0.0
cupshelpers==1.0
daphne==4.1.2
dbus-python==1.2.18
defer==1.0.6
deprecation==2.1.0
distlib==0.3.4
distro==1.9.0
dj-database-url==1.0.0
Django==5.0.7
django-cassandra-engine==1.9.0
djangorestframework==3.15.2
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
eyed3==0.9.7
filelock==3.15.4
filetype==1.2.0
geomet==0.2.1.post1
greenlet==3.0.3
gunicorn==20.1.0
httplib2==0.22.0
hyperlink==21.0.0
idna==3.7
ifaddr==0.2.0
IMDbPY==2022.7.9
importlib_metadata==8.2.0
incremental==24.7.2
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.2
jeepney==0.8.0
jsonschema==3.2.0
keyring==25.3.0
language-selector==0.1
launchpadlib==2.0.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
louis==3.20.0
lxml==5.2.2
macaroonbakery==1.3.4
Mako==1.3.5
Markdown==3.6
MarkupSafe==2.1.5
menulibre==2.2.2
more-itertools==10.3.0
mugshot==0.4.3
mutagen==1.47.0
natsort==8.4.0
netaddr==1.3.0
netifaces==0.11.0
numpy==2.0.1
oauthlib==3.2.2
onboard==1.4.1
packaging==24.1
pandas==2.2.3
pexpect==4.9.0
pillow==10.4.0
platformdirs==2.5.1
protobuf==5.27.3
psutil==6.0.0
psycopg2-binary==2.9.6
ptyprocess==0.7.0
py4j==0.10.9.7
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycairo==1.20.1
pycparser==2.22
pycryptodomex==3.20.0
pycups==2.0.1
pycurl==7.45.3
pyelftools==0.31
Pygments==2.18.0
PyGObject==3.42.1
PyICU==2.13.1
PyJWT==2.9.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==24.2.1
pyparsing==3.1.2
pyparted==3.11.7
pypng==0.20220715.0
pyRFC3339==1.1
pyrsistent==0.18.1
pyspark==3.5.3
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.49
python-dotenv==0.19.2
python-gnupg==0.5.2
python-magic==0.4.27
pytube==15.0.0
pytz==2024.1
pyxdg==0.28
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.2.2
requests==2.32.3
requests-file==2.1.0
screen-resolution-extra==0.0.0
scylla-driver==3.26.9
sdkmanager==0.5.1
SecretStorage==3.3.3
service-identity==24.2.0
setproctitle==1.3.3
six==1.16.0
soupsieve==2.5
speedtest-cli==2.1.3
SQLAlchemy==2.0.31
sqlparse==0.5.1
systemd-python==234
texttable==1.6.4
tldextract==5.1.2
toml==0.10.2
tomli==2.0.2
Twisted==24.10.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
ubuntu-drivers-common==0.0.0
ufw==0.36.1
Unidecode==1.3.8
urllib3==2.2.2
virtualenv==20.13.0+ds
wadllib==1.3.6
websocket-client==1.2.3
websockets==12.0
whitenoise==6.4.0
wxPython==4.0.7
xdg==6.0.0
xkit==0.0.0
yt-dlp==2024.8.1
zipp==3.19.2
zope.interface==7.1.1
